Refactoring the interference code
Simpler function that comput the velocity at each branch of the tree.
Now the wake midline gets stretched/compressed according to the velocity
induced by other wakes. Still not sure if this change is correct/accurate.

Certainly it is better than the constant length wake used previously.

I still need to apply the code refactorization to the wakeinterference function.