Did some simulations for CTFLORESTA
Very large numbers of branches fail to converge. I think the only way to do it is some recursive algorithm where the space is divided into sections and each section is substituted by a meta-branch. This case is converged and then we move on to higher resolutions.

For this to be done I will have to improve the code organization.